requirements.txt
Django==5.1.7
tweepy==4.14.0
kafka-python==2.0.2
pymongo==4.6.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
textblob==0.17.1
celery==5.3.6
redis==5.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-cors-headers==4.3.1
openpyxl==3.1.2
python-dateutil==2.8.2
django-celery-beat==2.5.0 